Pass parent object to connection during execution
The ability to create custom edge types was recently added and, with it,
the ability to pass a `parent` object through connections for use in the
connection and custom edge type. However, we forgot to actually _pass_
the parent object in during query execution, so it ended up always being
`nil`. This patch fixes the failing test introduced in the previous
commit.

Signed-off-by: David Celis <me@davidcel.is>