BuildFromDefinition: add parser option
This adds a callable `parser` option to `from_definition`.

The main use case for changing the default `GraphQL.parse` is if you're
using `graphql-libgraphqlparser-ruby`. That gem overrides `GraphQL.parse`
which does not support parsing an IDL string.

The workaround for this would be to pass in a callable that uses
`GraphQL.parse_without_libgraphqlparser`.